+++ test/os/lua_test.exs
@@ -80,6 +80,7 @@ defmodule FarmbotOS.LuaTest do
     "go_to_home(\"all\")",
     "go_to_home(\"x\")\ngo_to_home(\"y\")\ngo_to_home(\"z\")",
     "go_to_home()",
+    "group(1)",
     "move_absolute(1.0, 0, 0)\ncheck_position({x = 1.0, y = 0,  z = 0}, 0.50)",
     "move_absolute(1.0, 0, 0)\nmove_absolute(coordinate(1.0, 20, 30))",
     "move_absolute(20, 100, 100)\ncheck_position(coordinate(20, 100, 100), 1)",
@@ -93,6 +94,7 @@ defmodule FarmbotOS.LuaTest do
     "send_message(\"info\", \"Running FBOS v\" .. fbos_version())",
     "send_message(\"info\", \"Time zone is: \" .. get_device().timezone)",
     "send_message(\"info\", 23, {\"toast\"})",
+    "sort({}, \"random\")",
     "status = read_status()",
     "toggle_pin(13)",
     "update_device({name = \"Test Farmbot\"})",
@@ -160,6 +162,14 @@ defmodule FarmbotOS.LuaTest do
 
     expect(DataManipulation, :update_firmware_config, 1, fn
       [[{"encoder_enabled_z", 1.0}]], lua -> {[], lua}
+    end)
+
+    expect(FarmbotOS.Asset, :sort_points, 1, fn
+      [], "random" -> []
+    end)
+
+    expect(FarmbotOS.Asset, :find_points_via_group, 1, fn
+      1 -> %{point_ids: []}
     end)
 
     Enum.map(@documentation_examples, fn lua ->